+++ app/src/main/java/com/dessalines/thumbkey/keyboards/NumericWide.kt
@@ -31,7 +31,6 @@ val WIDE_NUMERIC_KEYBOARD =
                 KeyItemC(
                     center = KeyC("2", size = LARGE),
                     swipeType = FOUR_WAY_DIAGONAL,
-                    topLeft = KeyC("°"),
                     topRight =
                         KeyC(
                             // secret! (to avoid excessive visual noise)
@@ -42,6 +41,7 @@ val WIDE_NUMERIC_KEYBOARD =
                 KeyItemC(
                     center = KeyC("3", size = LARGE),
                     swipeType = FOUR_WAY_DIAGONAL,
+                    topLeft = KeyC("°"),
                     bottomLeft = KeyC("^"),
                     topRight =
                         KeyC(
@@ -55,7 +55,7 @@ val WIDE_NUMERIC_KEYBOARD =
             listOf(
                 KeyItemC(
                     center = KeyC("(", size = LARGE),
-                    swipeType = FOUR_WAY_CROSS,
+                    swipeType = EIGHT_WAY,
                     top = KeyC("<"),
                     topLeft = KeyC("{"),
                     left =
@@ -99,7 +99,7 @@ val WIDE_NUMERIC_KEYBOARD =
                 ),
                 KeyItemC(
                     center = KeyC(")", size = LARGE),
-                    swipeType = FOUR_WAY_CROSS,
+                    swipeType = EIGHT_WAY,
                     top = KeyC(">"),
                     topRight = KeyC("}"),
                     right =
@@ -124,7 +124,7 @@ val WIDE_NUMERIC_KEYBOARD =
                 ),
                 KeyItemC(
                     center = KeyC("7", size = LARGE),
-                    swipeType = FOUR_WAY_DIAGONAL,
+                    swipeType = EIGHT_WAY,
                     left = KeyC("@"),
                     top = KeyC("&"),
                     topRight = KeyC("%"),
@@ -135,13 +135,13 @@ val WIDE_NUMERIC_KEYBOARD =
                 KeyItemC(
                     center = KeyC("8", size = LARGE),
                     swipeType = FOUR_WAY_DIAGONAL,
-                    bottomLeft = KeyC("*"),
-                    bottomRight = KeyC("#"),
                 ),
                 KeyItemC(
                     center = KeyC("9", size = LARGE),
-                    swipeType = FOUR_WAY_DIAGONAL,
+                    swipeType = EIGHT_WAY,
                     top = KeyC("~"),
+                    left = KeyC("*"),
+                    right = KeyC("#"),
                     bottomLeft = KeyC("\\"),
                     bottomRight = KeyC("/"),
                 ),